requirements.txt
pyserial>=3.5
requests>=2.28
numpy>=1.20
scipy>=1.7
matplotlib>=3.5
